schemas: refactor jsonschemas compilation
* create ‘jsonschemas’ module

* updates way schemas should be constructed and used
