+++ cap/jsonschemas/deposits/common/deposit-v1.0.0.json
