+++ cap/jsonschemas/records/cms-analysis-v1.0.0-src.json
@@ -1,0 +1,21 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "type": "object",
+  "title": "Deposit schema.",
+  "description": "Describe information needed for deposit module.",
+  "properties": {
+    "basic_info": {
+      "type": "object",
+      "properties": {
+        "anal_name": {
+          "type": "string",
+          "title": "Analysis Name"
+        },
+        "cadi": {
+          "type": "string",
+          "title": "CADI #"
+        }
+      }
+    }
+  }
+}
