+++ cap/mappings/deposits/records/cms-analysis-v1.0.0.json
@@ -25,7 +25,7 @@
           "type": "object",
           "properties": {
             "id": {
-              "type": "string",
+              "type": "integer",
               "index": "not_analyzed"
             },
             "pid": {
