+++ cap/mappings/deposits/records/lhcb-v1.0.0.json
@@ -22,7 +22,7 @@
           "type": "object",
           "properties": {
             "id": {
-              "type": "string",
+              "type": "integer",
               "index": "not_analyzed"
             },
             "pid": {
