+++ cap/modules/jsonschemas/__init__.py
@@ -1,20 +1,20 @@
 # -*- coding: utf-8 -*-
 #
-# This file is part of CERN Analysis Preservation Framework.
+# This file is part of CERN Analysis Preservation.
 # Copyright (C) 2016 CERN.
 #
-# CERN Analysis Preservation Framework is free software; you can redistribute
-# it and/or modify it under the terms of the GNU General Public License as
+# CERN Analysis Preservation is free software; you can redistribute it
+# and/or modify it under the terms of the GNU General Public License as
 # published by the Free Software Foundation; either version 2 of the
 # License, or (at your option) any later version.
 #
-# CERN Analysis Preservation Framework is distributed in the hope that it will
-# be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
+# CERN Analysis Preservation is distributed in the hope that it will be
+# useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 # General Public License for more details.
 #
 # You should have received a copy of the GNU General Public License
-# along with CERN Analysis Preservation Framework; if not, write to the
+# along with CERN Analysis Preservation; if not, write to the
 # Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston,
 # MA 02111-1307, USA.
 #
@@ -22,12 +22,10 @@
 # waive the privileges and immunities granted to it by virtue of its status
 # as an Intergovernmental Organization or submit itself to any jurisdiction.
 
-"""cap base Invenio configuration."""
+"""Data model package."""
 
 from __future__ import absolute_import, print_function
 
-from invenio_base.app import create_cli
+from .ext import CAPJSONschemas
 
-from .factory import create_app
-
-cli = create_cli(create_app=create_app)
+__all__ = ('CAPJSONschemas', )
